show message box when deleting all config parameters and also delete modifications when config parameter is deleted

+++ src/db/controller/ConfigParameterController.ts
@@ -1,5 +1,7 @@
 import { getRepository } from 'typeorm';
 import { ConfigParameter } from '../entity/ConfigParameter';
+import { dialog } from 'electron';
+import { ParameterMod } from '../entity/ParameterMod';
 
 export class ConfigParameterController {
     public static getAll = async (): Promise<ConfigParameter[]> => {
@@ -36,10 +38,15 @@ export class ConfigParameterController {
 
     public static async deleteParameter(id: number): Promise<boolean> {
         const parameterRepo = getRepository(ConfigParameter);
+        const modRepo = getRepository(ParameterMod);
 
         return new Promise(async (resolve, reject) => {
             const parameter = await parameterRepo.findOne(id);
             if (parameter) {
+                const mods = await modRepo.find({ where: { configParameter: parameter } });
+                for (const mod of mods) {
+                    modRepo.delete(mod);
+                }
                 await parameterRepo.delete(id);
             }
 
@@ -50,11 +57,29 @@ export class ConfigParameterController {
 
     public static async deleteAll(): Promise<boolean> {
         const parameterRepo = getRepository(ConfigParameter);
+        const modRepo = getRepository(ParameterMod);
 
         return new Promise(async (resolve, reject) => {
-            const parameters = await parameterRepo.find();
-            for (const parameter of parameters) {
-                await parameterRepo.delete(parameter);
+            // show message box
+            const action = dialog.showMessageBox({
+                type: 'warning',
+                buttons: ['Ja', 'Abbrechen'],
+                defaultId: 0,
+                title: 'Alle Parameter löschen?',
+                message: 'Sollen wirklich alle Konfigurationsparameter gelöscht werden?',
+                cancelId: 1,
+            });
+
+            if (action === 0) {
+                const parameters = await parameterRepo.find();
+                for (const parameter of parameters) {
+                    const mods = await modRepo.find({ where: { configParameter: parameter } });
+                    for (const mod of mods) {
+                        await modRepo.delete(mod);
+                    }
+                    await parameterRepo.delete(parameter);
+                }
+
             }
 
             // resolve promise when done
